Custom sort for Start/End. Adding Contemporary period. DataGrid binding.
Datagrid is now bound to a list in the model, everytime the list is updated, the datagrid is updated. This means we do not need to add manually and remove from the datagrid ourselves. This also means we do not need to create the columns ourselves but we still need to edit them once they're created.
Using a custom sort (comparer) for the Start and End columns since they are stored as strings and use values such as Present to represent right now. It was not being sorted properly since it was not ints.